May 27
So, I have been working on map generator and now we have an user interface map generator for the game. the project have an incomplete intro screen that has three buttons. one direct to map generator, and the other switch to map loader for the user to play their own map generated level. There are some colliders on the map that are drawn by a loop. the loop identifies the file path of the images and then draws a collider where the img file has been places on the map. the map is no more a big png file, but  instead it's a grid with 280 small png files that make the whole map. the map file that can be store is no more a picture, but a object(map) that stores two layers of base and decoration. and the base png files are divided into two folders the ones that are used for the background and others for borders. the collider loop basically put a collider on each border png files. however, the decoration files are not separated. there are some colliders that are modified for certain png files. for the king.js we were using one hit box for movement, but it caused the character to climb walls as it thought it touching ground. (all walls(colliders) have been created in a group and all been called walls), so I added a new hit box under the character that act as touching the ground instead of identifying which collider is the ground border.