End of Class June 2 and the weekend before
Fixed the Pig Problems so now all pigs are same size, and start to follow the player when the distance between them crosses a threshold.

Also fixed the jump problems. so used a sensor approach to do that.

Also fixed the jitters in the animations.

and now the intro screen has a background of our intro screen.