End of Class Jun 6
Added the pig knockback.
So now the pigs get knocked back when attackde by the king, also added the side check for the attack so the king also can only attack the side it is facing.

And laslty trying something new so the  when the king jumps on top of the pig the pig takes 1 damage and the king jumps high.